+++ types/api.go
@@ -67,7 +67,7 @@ type MetaResponse struct {
 	Files                []byte     `json:"files,omitempty"`
 	Finished             *bool      `json:"finished,omitempty"`
 	Health               int32      `json:"health,omitempty"`
-	Height               *int64     `json:"height,omitempty"`
+	Height               *uint32    `json:"height,omitempty"`
 	InvalidClaims        []int      `json:"invalidClaims,omitempty"`
 	Message              string     `json:"message,omitempty"`
 	Nullifier            string     `json:"nullifier,omitempty"`
